Refactor creating BugRow component
