+++ src/components/BugRow/BugRow.scss
@@ -1,0 +1,17 @@
+.bug {
+  &__image {
+    border-radius: 50%;
+    width: 70px;
+    aspect-ratio: 1;
+  }
+
+  &__name {
+    font-weight: bold;
+  }
+
+  &__actions {
+    & > :not(:first-child).button {
+      margin-left: 50px;
+    }
+  }
+}
